Fix REPL issue, restructure standard functions
- REPL issue:
  - If an exception was thrown, the thread would not be removed, leading to guaranteed rethrow of the same exception.

- Restructure standard functions
  - Move to stdlib folder, start splitting out by functionality
