+++ PocoLithp/ELisp.cpp
@@ -5,5 +5,8 @@
 namespace Elispidae {
 	void init_runtime() {
 		Threads::init_threads();
+		Stdlib::init_dictionary();
+		Stdlib::init_lists();
+		Stdlib::init_strings();
 	}
 }
