+++ PocoLithp/stdlib/ELthreads.cpp
@@ -1,8 +1,5 @@
 #include "stdafx.h"
 
-#include "PLint_stackless.hpp"
-#include "ELisp.hpp"
-
 using namespace PocoLithp;
 using namespace PocoLithp::Stackless;
