+++ PocoLithp/stdlib/stdafx.h
@@ -1,0 +1,7 @@
+#pragma once
+
+// TODO: This ugly hack is needed to resolve issues in Visual Studio.
+// Intellisense says that project symbols are missing, but compiles fine.
+// https://stackoverflow.com/questions/31943634/visual-studio-2015-intellisense-errors-but-solution-compiles
+
+#include "../stdafx.h"
